backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.3
daphne==4.1.2
Django==4.2.16
django-allauth==65.0.2
django-celery-beat==2.7.0
django-cleanup==9.0.0
django-cors-headers==4.4.0
django-htmx==1.19.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-prometheus==2.1.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
factory_boy==3.3.3
Faker==30.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.4.2
msgpack==1.1.0
numpy==2.2.5
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
pillow==11.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
rpds-py==0.24.0
selenium==4.29.0
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tomli==2.2.1
toposort==1.10
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2
